3-client-socket-fork-구현

+++ head/common.h
@@ -1,5 +1,7 @@
 #pragma once
 
+
+#include <iostream>
 #include <string>
 #include <regex>
 
@@ -28,7 +30,7 @@
 #define RTU_STATUS_REQ (unsigned char)0x19
 #define RTU_STATUS_RES (unsigned char)0x1F
 
-typedef size_t SOCKET;
+typedef int SOCKET;
 
 namespace core {
     namespace common {
